feat: Docker config
 - Eslint setup
 - VSCode launch setup
 - Remove dotenv and use docker env variables
 - Add README
 - Run express add using http createServer
