+++ Dockerfile
@@ -1,0 +1,11 @@
+FROM buildkite/puppeteer
+
+WORKDIR /usr/src/app
+
+COPY ./package.json .
+
+RUN npm install --only=prod
+
+COPY . .
+
+CMD ["npm", "start"]
