+++ README.md
@@ -1,0 +1,7 @@
+## Setup
+* Set all the missing env variables in docker-compose.yml
+
+## Run in development
+```
+  docker-compose up --build
+```
