+++ app/funcionalities/commonHours.js
