+++ app/funcionalities/messageClass.js
