+++ db/mongo-init.js
@@ -1,0 +1,10 @@
+db.createUser({
+  user: 'holu_user',
+  pwd: 'root_12345',
+  roles: [
+    {
+      role: 'readWrite',
+      db: 'quevent',
+    },
+  ],
+});
